functions/daily_summary/requirements.txt
google-cloud-bigquery>=3.27.0,<4.0
